simplify specialization and optimize bounds check for nulls
