+++ crates/polars-utils/src/index.rs
@@ -3,14 +3,18 @@ use polars_error::{polars_bail, polars_ensure, PolarsResult};
 use crate::IdxSize;
 
 pub fn check_bounds(idx: &[IdxSize], len: IdxSize) -> PolarsResult<()> {
-    let mut inbounds = true;
-
-    for &i in idx {
-        if i >= len {
-            // we will not break here as that prevents SIMD
-            inbounds = false;
+    // We iterate in large uninterrupted chunks to help auto-vectorization.
+    let mut in_bounds = true;
+    for chunk in idx.chunks(1024) {
+        for i in chunk {
+            if *i >= len {
+                in_bounds = false;
+            }
+        }
+        if !in_bounds {
+            break;
         }
     }
-    polars_ensure!(inbounds, ComputeError: "indices are out of bounds");
+    polars_ensure!(in_bounds, ComputeError: "indices are out of bounds");
     Ok(())
 }
